requirements.txt
annotated-types==0.7.0
apache-beam==2.62.0
attrs==25.1.0
cachetools==4.2.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
dill==0.3.1.1
fastavro==1.10.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-aiplatform==1.80.0
google-cloud-bigquery==3.29.0
google-cloud-storage==2.19.0
grpcio==1.65.5
hdfs==2.7.3
httplib2==0.20.4
idna==3.10
joblib==1.4.2
nltk==3.9.1
numpy==1.26.0
pandas==2.0.3
protobuf==3.20.3
pyarrow==9.0.0
pydantic==2.10.6
pymongo==3.13.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.11.4
six==1.17.0
tqdm==4.67.1
urllib3==2.3.0